+++ traffic_ops/testing/api/v2/monitoring_test.go
@@ -42,8 +42,8 @@ func AllCDNsCanSnapshot(t *testing.T) {
 			continue
 		}
 
-		_, _, err = TOSession.GetTrafficMonitorConfigMap(cdn.Name)
-		if err != nil {
+		tmConfig, _, err := TOSession.GetTrafficMonitorConfigMap(cdn.Name)
+		if err != nil && tmConfig == nil {
 			t.Error(err)
 			continue
 		}