+++ traffic_ops_ort/atstccfg/config/config.go
@@ -236,7 +236,7 @@ func (fs ATSConfigFiles) Swap(i, j int) { fs[i], fs[j] = fs[j], fs[i] }
 //   require the potential fields to be omitted to generate correctly.
 type TOData struct {
 	// Servers must be all the servers from Traffic Ops. May include servers not on the current cdn.
-	Servers []tc.Server
+	Servers []tc.ServerNullable
 
 	// CacheGroups must be all cachegroups in Traffic Ops with Servers on the current server's cdn. May also include CacheGroups without servers on the current cdn.
 	CacheGroups []tc.CacheGroupNullable
@@ -263,7 +263,7 @@ type TOData struct {
 	DeliveryServiceServers []tc.DeliveryServiceServer
 
 	// Server must be the server we're fetching configs from
-	Server tc.Server
+	Server *tc.ServerNullable
 
 	// TOToolName must be the Parameter named 'tm.toolname' on the tc.GlobalConfigFileName Profile.
 	TOToolName string