+++ traffic_portal/app/src/modules/private/cacheGroups/servers/index.js
@@ -30,8 +30,17 @@ module.exports = angular.module('trafficPortal.private.cacheGroups.servers', [])
 							cacheGroup: function($stateParams, cacheGroupService) {
 								return cacheGroupService.getCacheGroup($stateParams.cacheGroupId);
 							},
-							servers: function($stateParams, serverService) {
-								return serverService.getServers({ cachegroup: $stateParams.cacheGroupId, orderby: 'hostName' });
+							servers: function(cacheGroup, $stateParams, serverService) {
+								return serverService.getServers({ cachegroup: cacheGroup.id, orderby: 'hostName' });
+							},
+							filter: function(cacheGroup) {
+								return {
+									cachegroup: {
+										filterType: "text",
+										type: "equals",
+										filter: cacheGroup.name
+									}
+								}
 							}
 						}
 					}