+++ traffic_portal/app/src/modules/private/profiles/servers/index.js
@@ -30,8 +30,17 @@ module.exports = angular.module('trafficPortal.private.profiles.servers', [])
 							profile: function($stateParams, profileService) {
 								return profileService.getProfile($stateParams.profileId);
 							},
-							servers: function($stateParams, serverService) {
-								return serverService.getServers({ profileId: $stateParams.profileId, orderby: 'hostName' });
+							servers: function(profile, $stateParams, serverService) {
+								return serverService.getServers({ profileId: profile.id, orderby: 'hostName' });
+							},
+							filter: function(profile) {
+								return {
+									profile: {
+										filterType: "text",
+										type: "equals",
+										filter: profile.name
+									}
+								}
 							}
 						}
 					}